Added support for sizeLimit parameter in sources.xml.
git-svn-id: http://unused-160-33.aus.redhat.com/svn/penrose-server/branches/1.2@884 04c32f3d-caf7-0310-9fe0-d49f5cee804d

+++ core/src/java/org/safehaus/penrose/ldap/SearchResponse.java
@@ -39,7 +39,7 @@ public class SearchResponse<E> extends Response {
 
     protected long bufferSize;
     protected long sizeLimit;
-    protected int totalCount;
+    protected long totalCount;
     protected boolean closed = false;
 
     protected LDAPException exception;
@@ -262,7 +262,7 @@ public class SearchResponse<E> extends Response {
         notifyAll();
     }
 
-    public int getTotalCount() throws Exception {
+    public long getTotalCount() throws Exception {
         return totalCount;
     }
