+++ core/src/test/org/safehaus/penrose/test/mapping/nested2/SearchBaseTest.java
@@ -55,7 +55,7 @@ public class SearchBaseTest extends NestedTestCase {
         log.debug("hasNext: "+hasNext);
         assertFalse(hasNext);
 
-        int totalCount = response.getTotalCount();
+        long totalCount = response.getTotalCount();
         log.debug("totalCount: "+totalCount);
         assertEquals(1, totalCount);
 
@@ -104,7 +104,7 @@ public class SearchBaseTest extends NestedTestCase {
         log.debug("hasNext: "+hasNext);
         assertFalse(hasNext);
 
-        int totalCount = response.getTotalCount();
+        long totalCount = response.getTotalCount();
         log.debug("totalCount: "+totalCount);
         assertEquals(1, totalCount);
